fix: use next link and move images to public folder
